# openapi.json
{
  "openapi": "3.0.0",
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Development — بيئة التطوير"
    },
    {
      "url": "https://akfeek-backend.developteam.site",
      "description": "Production — بيئة الإنتاج"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT token — أدخل التوكن بعد كلمة Bearer مثال: `Bearer eyJhbGci...`"
      }
    },
    "schemas": {
...
      "Success": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean",
            "example": true
          },
          "message": {
            "type": "string",
            "example": "Operation successful"
          },
          "messageAr": {
            "type": "string",
            "example": "تمت العملية بنجاح"
          }
        }
      },
      "Pagination": {
        "type": "object",
        "properties": {
          "page": {
            "type": "integer",
            "example": 1
          },
          "limit": {
            "type": "integer",
            "example": 10
          },
          "total": {
            "type": "integer",
            "example": 100
          },
          "totalPages": {
            "type": "integer",
            "example": 10
          }
        }
      },
...
      "VendorProfile": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "userId": {
            "type": "string",
            "format": "uuid"
          },
          "businessName": {
            "type": "string"
          },
          "businessNameAr": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "descriptionAr": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "PENDING_APPROVAL",
              "ACTIVE",
              "SUSPENDED",
              "REJECTED"
            ]
          },
          "isActive": {
            "type": "boolean"
          },
          "logo": {
            "type": "string"
          },
          "contactEmail": {
            "type": "string"
          },
          "contactPhone": {
            "type": "string"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
...
    },
    "parameters": {
...
      "PageParam": {
        "name": "page",
        "in": "query",
        "description": "رقم الصفحة | Page number",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "default": 1
        }
      },
      "LimitParam": {
        "name": "limit",
        "in": "query",
        "description": "عدد العناصر في الصفحة | Items per page",
        "schema": {
          "type": "integer",
          "minimum": 1,
          "maximum": 100,
          "default": 10
        }
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "info": {
    "title": "🚗 Akfeek API — Full Reference",
    "version": "1.0.0",
    "description": "\n# Akfeek Platform — Full API Reference\n\n> **للمطورين الداخليين فقط** — هذه الوثيقة تحتوي على **جميع** الـ endpoints.\n> لو بتشتغل على جزء معين، استخدم الرابط المخصص ليك:\n\n| الجمهور | الرابط |\n|---|---|\n| 📱 تطبيق موبايل العميل | [`/api-docs/mobile`](/api-docs/mobile) |\n| 🔧 تطبيق موبايل الفني | [`/api-docs/technician`](/api-docs/technician) |\n| 🏪 بوابة الفيندور / الورشة | [`/api-docs/vendor`](/api-docs/vendor) |\n| ⚙️ لوحة تحكم الأدمن | [`/api-docs/admin`](/api-docs/admin) |\n\n---\n\n## Authentication / المصادقة\n\n```\nAuthorization: Bearer <your_jwt_token>\n```\n\n1. `POST /api/auth/login` — سجل دخول واحصل على التوكن\n2. انقر **Authorize** في الأعلى وأدخل: `Bearer <token>`\n      ",
    "contact": {
      "name": "Akfeek API Support",
      "email": "support@akfeek.com"
    }
  },
...
  "paths": {
...
    "/api/auto-parts/vendor/{vendorId}": {
      "get": {
        "summary": "Get parts by vendor",
        "tags": [
          "🏪 Vendor | Spare Parts"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "vendorId",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of vendor parts"
          }
        }
      }
    },
...
    "/api/vendors": {
      "get": {
        "summary": "Get all vendors (Admin only)",
        "tags": [
          "📱 Customer | Vendors"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string"
            },
            "description": "Filter by vendor status"
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            },
            "description": "Search by business name"
          },
          {
            "in": "query",
            "name": "vendorType",
            "schema": {
              "type": "string",
              "enum": [
                "AUTO_PARTS",
                "COMPREHENSIVE_CARE",
                "CERTIFIED_WORKSHOP",
                "CAR_WASH",
                "ADHMN_AKFEEK"
              ]
            }
          },
          {
            "in": "query",
            "name": "isVerified",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "$ref": "#/components/parameters/PageParam"
          },
          {
            "$ref": "#/components/parameters/LimitParam"
          }
        ],
        "responses": {
          "200": {
            "description": "List of vendors",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/VendorProfile"
                      }
                    },
                    "pagination": {
                      "$ref": "#/components/schemas/Pagination"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "summary": "Create new vendor profile",
        "tags": [
          "📱 Customer | Vendors"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "businessName",
                  "contactPhone"
                ],
                "properties": {
                  "userId": {
                    "type": "string",
                    "description": "User ID (Required for Admin, ignored for self-registration)"
                  },
                  "businessName": {
                    "type": "string"
                  },
                  "businessNameAr": {
                    "type": "string"
                  },
                  "contactEmail": {
                    "type": "string"
                  },
                  "contactPhone": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Vendor created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Success"
                }
              }
            }
          }
        }
      }
    },
    "/api/vendors/coupons": {
      "get": {
        "summary": "Get all vendor coupons (Admin Only)",
        "description": "Retrieve all coupons across all vendors with optional filtering.",
        "tags": [
          "🏪 Vendor | Workshop"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "isActive",
            "schema": {
              "type": "boolean"
            }
          },
          {
            "in": "query",
            "name": "vendorId",
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "search",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of all coupons"
          }
        }
      }
    },
    "/api/vendors/profile/me": {
      "get": {
        "summary": "Get current user's vendor profile",
        "tags": [
          "📱 Customer | Vendors"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current vendor profile",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/VendorProfile"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/vendors/profile/me/comprehensive-care-bookings": {
      "get": {
        "summary": "Get comprehensive care service bookings for current vendor",
        "tags": [
          "🏪 Vendor | Comprehensive Care"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of bookings"
          }
        }
      }
    },
    "/api/vendors/profile/me/coupons": {
      "get": {
        "summary": "Get all coupons for current vendor",
        "tags": [
          "🏪 Vendor | Workshop"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "List of vendor coupons"
          }
        }
      },
      "post": {
        "summary": "Create new coupon for current vendor",
        "tags": [
          "🏪 Vendor | Workshop"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "code",
                  "discountType",
                  "discountValue"
                ],
                "properties": {
                  "code": {
                    "type": "string"
                  },
                  "discountType": {
                    "type": "string",
                    "enum": [
                      "PERCENTAGE",
                      "FIXED"
                    ]
                  },
                  "discountValue": {
                    "type": "number"
                  },
                  "minOrderValue": {
                    "type": "number"
                  },
                  "startDate": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "endDate": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "usageLimit": {
                    "type": "integer"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Coupon created"
          }
        }
      }
    },
    "/api/vendors/profile/me/coupons/{id}": {
      "patch": {
        "summary": "Update vendor coupon",
        "tags": [
          "🏪 Vendor | Workshop"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Coupon updated"
          }
        }
      },
      "delete": {
        "summary": "Delete vendor coupon",
        "tags": [
          "🏪 Vendor | Workshop"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Coupon deleted"
          }
        }
      }
    },
    "/api/vendors/{id}": {
      "get": {
        "summary": "Get vendor details by ID",
        "tags": [
          "📱 Customer | Vendors"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Vendor details",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean"
                    },
                    "data": {
                      "$ref": "#/components/schemas/VendorProfile"
                    }
                  }
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update vendor profile",
        "tags": [
          "📱 Customer | Vendors"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "businessName": {
                    "type": "string"
                  },
                  "description": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Vendor updated"
          }
        }
      },
      "delete": {
        "summary": "Delete vendor (Admin only)",
        "tags": [
          "📱 Customer | Vendors"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Vendor deleted"
          }
        }
      }
    },
    "/api/vendors/{id}/stats": {
      "get": {
        "summary": "Get vendor statistics",
        "tags": [
          "📱 Customer | Vendors"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Vendor statistics"
          }
        }
      }
    },
    "/api/vendors/{id}/reviews": {
      "get": {
        "summary": "Get vendor reviews",
        "tags": [
          "📱 Customer | Vendors"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of vendor reviews"
          }
        }
      },
      "post": {
        "summary": "Submit a review for a vendor",
        "tags": [
          "📱 Customer | Vendors"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "rating",
                  "comment"
                ],
                "properties": {
                  "rating": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 5
                  },
                  "comment": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Review submitted"
          }
        }
      }
    },
    "/api/vendors/{id}/status": {
      "put": {
        "summary": "Update vendor status (Admin only)",
        "tags": [
          "📱 Customer | Vendors"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "status"
                ],
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": [
                      "PENDING_APPROVAL",
                      "ACTIVE",
                      "SUSPENDED",
                      "REJECTED"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status updated"
          }
        }
      }
    },
...
  }
}